+++ test/DependencyInjection/Loader/JsFileLoader.spec.js
@@ -41,6 +41,7 @@ describe('JsFileLoader', () => {
       loader.load()
       let service = container.get(serviceName)
       let aliasService = container.get(aliasName)
+      let taggedServices = container.findTaggedServiceIds('fooTag')
 
       // Assert.
       assert.instanceOf(service, Foo)
@@ -49,6 +50,7 @@ describe('JsFileLoader', () => {
       assert.isFunction(service.fs.copy)
       assert.strictEqual(service.param, 'foo-bar')
       assert.strictEqual(aliasService, service)
+      assert.lengthOf(taggedServices.toArray(), 2)
     })
   })
 })